backend/requirements.txt
django>=5.2.5
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1,
djoser>=2.3.3,
django-extensions>=4.1,
pillow>=11.3.0,
drf-extra-fields>=3.7.0,
django-filter>=25.1,
gunicorn